simplified more betterer smooth move
For some reason I thought it was a good idea to put render in a loop and break out of it if it's time to calc the next tic. But, you know, like... no, man. And if I think about it a bit more, maybe I can only call rl.GetTime() once per loop.