more betterer smooth motion II: the bettererening
Both time trackers for calc time and draw time were keeping track of the time of the entire game_loop instead of how much time it takes to calc and how much time it takes to draw. The game loop is artificially inflated to 60FPS, so it thought it was taking way too long to do a few measly calculations and skipping draw calls to compensate. And I've just realised a potential bug if calc ever DOES take too long. It will still be using the last draw time to figure its own calc time, so... I'll fix that.